requirements.txt
absl-py==2.0.0
astunparse==1.6.3
Babel==2.13.1
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.3.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.10.0
Cython==3.0.3
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.9.2
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-measurement-protocol==1.1.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lapx==0.5.5
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
prices==1.1.1
protobuf==3.19.6
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.7
rsa==4.9
scikit-video==1.1.11
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
SoccerNet==0.1.53
supervision==0.16.0
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thop==0.1.1.post2209072238
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.198
urllib3==2.0.6
werkzeug==3.0.0
wrapt==1.15.0
zipp==3.17.0
